Reorganize interfaces/components into separate files/subfolders
