+++ src/components/Experience/TimelineItem/TimelineItem.interface.ts
@@ -1,0 +1,8 @@
+export interface TimelineItemProps {
+  gridPositionClasses?: string;
+  company: string;
+  position: string;
+  isCurrentJob?: boolean;
+  startYear: string;
+  endYear?: string;
+}
