+++ src/components/Experience/TimelineItem/index.ts
@@ -1,0 +1,3 @@
+import TimelineItem from "./TimelineItem";
+
+export default TimelineItem;
