+++ src/components/Footer/SocialIcon/SocialIcon.interface.ts
@@ -1,0 +1,5 @@
+export interface SocialIconProps {
+  link: string;
+  imageSrc: string;
+  altName: string;
+}
