+++ src/components/Footer/SocialIcon/index.ts
@@ -1,0 +1,3 @@
+import SocialIcon from "./SocialIcon";
+
+export default SocialIcon;
