+++ src/components/Navbar/HamburgerButton/HamburgerButton.interface.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from "react";
+
+export interface HamburgerButtonProps {
+  isNavbarExpanded: boolean;
+  setIsNavbarExpanded: Dispatch<SetStateAction<boolean>>;
+}
