+++ src/components/Navbar/HamburgerButton/index.ts
@@ -1,0 +1,3 @@
+import HamburgerButton from "./HamburgerButton";
+
+export default HamburgerButton;
