+++ src/components/Navbar/MobileNavbar/index.ts
@@ -1,0 +1,3 @@
+import MobileNavbar from "./MobileNavbar";
+
+export default MobileNavbar;
