+++ src/components/Navbar/ThemeSelector/ThemeSelector.interface.ts
@@ -1,0 +1,12 @@
+import { Dispatch, SetStateAction } from "react";
+
+export interface ThemeSelectorOptionProps {
+  materialIcon: string;
+  label: string;
+  currentTheme: string;
+  setCurrentTheme: Dispatch<SetStateAction<string>>;
+}
+
+export interface ThemeSelectorProps {
+  isNavbarExpanded: boolean;
+}
