+++ src/components/Navbar/ThemeSelector/index.ts
@@ -1,0 +1,3 @@
+import ThemeSelector from "./ThemeSelector";
+
+export default ThemeSelector;
