+++ src/components/Projects/ProjectCard/ProjectCard.interface.ts
@@ -1,0 +1,7 @@
+import { GithubProject } from "../Projects.interface";
+
+export interface ProjectCardProps {
+  project: GithubProject;
+  type: "Recent" | "Starred";
+  loading: boolean;
+}
