+++ src/components/SkillsAndInterests/InterestAreaCard/InterestAreaCard.interface.ts
@@ -1,0 +1,5 @@
+export interface InterestAreaCardProps {
+  icon: string;
+  title: string;
+  description: string;
+}
