+++ src/components/SkillsAndInterests/InterestAreaCard/index.ts
@@ -1,0 +1,3 @@
+import InterestAreaCard from "./InterestAreaCard";
+
+export default InterestAreaCard;
