+++ src/components/SkillsAndInterests/SkillIcon/SkillIcon.interface.ts
@@ -1,0 +1,7 @@
+export interface SkillIconProps {
+  name: string;
+  type: "plain" | "original" | "plain-wordmark" | "original-wordmark";
+  size?: string;
+  caption?: string;
+  isColored?: boolean;
+}
