+++ src/components/SkillsAndInterests/SkillIcon/index.ts
@@ -1,0 +1,3 @@
+import SkillIcon from "./SkillIcon";
+
+export default SkillIcon;
