+++ src/components/SkillsAndInterests/SkillRow/SkillRow.enum.ts
@@ -1,0 +1,5 @@
+export enum SkillSetOptions {
+  DELVING_IN = "Languages & Tools I'm delving into",
+  DAILY_BASIS = "What I use on a daily basis",
+  OTHER_SKILLS = "Other Languages & Tools I'm skilled in",
+}
