+++ src/components/SkillsAndInterests/SkillRow/SkillRow.interface.ts
@@ -1,0 +1,3 @@
+export interface SkillRowProps {
+  selectedSkillSet: string;
+}
