+++ src/components/SkillsAndInterests/SkillRow/index.ts
@@ -1,0 +1,3 @@
+import SkillRow from "./SkillRow";
+
+export default SkillRow;
