+++ src/components/SkillsAndInterests/SkillRowSelector/SkillRowSelector.interface.ts
@@ -1,0 +1,6 @@
+import { Dispatch } from "react";
+
+export interface SkillRowSelectorProps {
+  selectedSkillSet: string;
+  onSelectSkillSet: Dispatch<string>;
+}
