+++ src/components/SkillsAndInterests/SkillRowSelector/index.ts
@@ -1,0 +1,3 @@
+import SkillRowSelector from "./SkillRowSelector";
+
+export default SkillRowSelector;
