+++ src/components/ThemeProvider/ThemeProvider.interface.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from "react";
+
+export interface IThemeContext {
+  currentTheme: string;
+  setCurrentTheme: Dispatch<SetStateAction<string>>;
+}
